payment service transactions edit

+++ payment-service/src/main/java/com/dp/paymentservice/model/transaction/TransactionWithdrawRequest.java
@@ -1,7 +1,9 @@
 package com.dp.paymentservice.model.transaction;
 
+import java.math.BigDecimal;
+
 public record TransactionWithdrawRequest (
         Long userId,
-        double amount
+        BigDecimal amount
 ) {
 }
